+++ src/App.tsx
@@ -1,10 +1,16 @@
 import {AppShell} from '@components/layout/AppShell'
+import {useTauriMenu} from '@hooks/useTauriMenu'
 import {AppProvider} from '@store/context'
+
+function AppInner() {
+    useTauriMenu()
+    return <AppShell/>
+}
 
 export default function App() {
     return (
         <AppProvider>
-            <AppShell/>
+            <AppInner/>
         </AppProvider>
     )
 }